Fix signWithWSSEInfo concurrency problem
When signWithWSSEInfo is passed the same WSSEAuthInfo object
on different threads, it's possible for the ID fields to change
while the sign method is executing resulting in a payload that
isn't signed correctly.

This change splits out the generated IDs from WSSEAuthInfo into
their own var so that they aren't part of the state shared
between calls.